Add NoteViewPhrase class and tests
